drinks

+++ drinks/app/Http/Controllers/TypeController.php
@@ -94,8 +94,15 @@ class TypeController extends Controller
      */
     public function destroy(Type $type)
     {
-        // dar grisim cia
-        $type->delete();
-        return redirect()->route('types-index');
+        if (!$type->typeDrinks()->count()) {
+            $type->delete();
+            return redirect()->route('types-index');
+        }
+        return 'negalima';
+        
+
+        
+        
+        
     }
 }
